create new file for api functions
