+++ go.mod
@@ -2,4 +2,4 @@ module github.com/rranar/telegram-api-webapp
 
 go 1.15
 
-require github.com/joho/godotenv v1.3.0 // indirect
+require github.com/joho/godotenv v1.3.0
