+++ main.go
@@ -1,78 +1,23 @@
 package main
 
 import (
-	"encoding/json"
 	"fmt"
-	"io/ioutil"
 	"log"
 	"net/http"
-	"net/url"
-	"os"
-	"strconv"
 	"strings"
 
-	"github.com/joho/godotenv"
 	"github.com/rranar/telegram-api-webapp/tgbot_core"
 )
 
-func parseTelegramRequest(r *http.Request) (*tgbot_core.Update, error) {
-	var update tgbot_core.Update
-	if err := json.NewDecoder(r.Body).Decode(&update); err != nil {
-		log.Println("Cannot decode incoming update %s", err.Error())
-		return nil, err
-	}
-	return &update, nil
-}
-
-func sendTextToTelegramChat(chatId int, text string) (string, error) {
-	log.Printf("Sending %s to chat id:%d", text, chatId)
-	var telegramToken, tokenError = getTelegramToken()
-	if tokenError != nil {
-		log.Printf("Something wrong with token - %s", tokenError.Error())
-		return "", tokenError
-	}
-
-	var telegramApi string = "https://api.telegram.org/bot" + telegramToken + "/sendMessage"
-	response, err := http.PostForm(
-		telegramApi,
-		url.Values{
-			"chat_id": {strconv.Itoa(chatId)},
-			"text":    {text},
-		})
-	if err != nil {
-		log.Printf("error when posting text to the chat:%s", err.Error())
-		return "", err
-	}
-	defer response.Body.Close()
-
-	var bodyBytes, errRead = ioutil.ReadAll(response.Body)
-	if errRead != nil {
-		log.Printf("error is parsing telegram answer:%s", errRead.Error())
-		return "", errRead
-	}
-	bodyString := string(bodyBytes)
-	log.Printf("Body of Telegram Response: %s", bodyString)
-
-	return bodyString, nil
-}
-
-func getTelegramToken() (string, error) {
-	err := godotenv.Load()
-	if err != nil {
-		return "", err
-	}
-	return os.Getenv("TELEGRAM_BOT_TOKEN"), nil
-}
-
 func HandlerTelegramWebhook(w http.ResponseWriter, r *http.Request) {
-	var update, err = parseTelegramRequest(r)
+	var update, err = tgbot_core.ParseTelegramRequest(r)
 	if err != nil {
 		log.Printf("error parsing update, %s", err.Error())
 		return
 	}
 
 	if strings.Contains(update.Message.Text, "moroz") {
-		var telegramResponseBody, errTelegram = sendTextToTelegramChat(update.Message.Chat.Id, "Морозило бродяга!Выгони его из хаты!")
+		var telegramResponseBody, errTelegram = tgbot_core.SendTextToTelegramChat(update.Message.Chat.Id, "Морозило бродяга!Выгони его из хаты!")
 		if errTelegram != nil {
 			log.Printf("got error from Telegram, response body is %s", errTelegram.Error(), telegramResponseBody)
 		}
